requirements.txt
numpy
scipy
scikit-image
opencv-python
tensorflow-gpu
tensorflow-model-optimization
matplotlib
tqdm
pillow
mnn
